Safe casting for fmpz

+++ fmpz/divexact_si.c
@@ -42,7 +42,7 @@ void fmpz_divexact_si(fmpz_t f, const fmpz_t g, slong h)
         }
         else
         {
-            flint_mpz_divexact_ui(mf, COEFF_TO_PTR(c1), -h);
+            flint_mpz_divexact_ui(mf, COEFF_TO_PTR(c1), -((ulong) h));
             _fmpz_demote_val(f);  /* division by h may result in small value */
             fmpz_neg(f, f);
         }
